FileTypeプロトコルを作成。Imageに適用した
